backend/documents/requirements.txt
psycopg2-binary>=2.9.9
boto3>=1.34.0
PyJWT>=2.8.0
